+++ BailarinaPreparadaApp/Models/Evaluations/EvaluationExercise.cs
@@ -10,7 +10,7 @@ namespace BailarinaPreparadaApp.Models.Evaluations
         public int ExerciseId { get; set; }
         public Exercise Exercise { get; set; } = null!;
         public ExerciseSide Side { get; set; } = ExerciseSide.None;
-        public int Score { get; set; } = 0;
+        public decimal Score { get; set; } = 0;
         public string Observation { get; set; } = string.Empty;
     }
 }
